feat(home): Add cards with summary of gathered data

+++ project/src/main/java/net/piotrl/music/modules/lastfm/track/repository/ScrobbleCrudRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface ScrobbleCrudRepository extends CrudRepository<ScrobbleEntity, Integer> {
 
     List<ScrobbleEntity> findAllByAccountId(long accountId);
-    long countByAccountId(long accountId);
+    int countByAccountId(long accountId);
 
 
 }
